requirements.txt
streamlit
opencv-python
numpy
tensorflow
Pillow
matplotlib
pandas
seaborn
scikit-learn